Add inline documentation to Sandbox and move GuestVariant to own header
