+++ src/gvar.cpp
@@ -1,4 +1,4 @@
-#include "sandbox.h"
+#include "guest_datatypes.h"
 
 #include <godot_cpp/variant/utility_functions.hpp>
 #include <libriscv/util/crc32.hpp>
